feat(01): complete phase 1 modernization and optimization
- Upgrade core stack: Laravel 12, Livewire 4, Tailwind CSS 4, and daisyUI 5.
- Implement strict type hints and return type declarations across models, forms, and controllers.
- Optimize QR Scanner with camera prioritization and hardware integration.
- Harden security with redundant authorization and strict validation.
- Implement mobile expansion features including PWA support and Hybrid Bottom Navigation.
- Clean up legacy Tailwind migration artifacts and modernize layout structure.

+++ database/factories/AttendanceFactory.php
@@ -9,7 +9,7 @@ use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Attendance>
+ * @extends Factory<Attendance>
  */
 class AttendanceFactory extends Factory
 {
@@ -27,7 +27,7 @@ class AttendanceFactory extends Factory
 
         $durationMinutes = null;
         if ($timeOut) {
-            $durationMinutes = Carbon::parse($timeIn)->diffInMinutes(Carbon::parse($timeOut));
+            $durationMinutes = (int) Carbon::parse($timeIn)->diffInMinutes(Carbon::parse($timeOut));
         }
 
         $status = $timeOut ? 'completed' : 'active';
@@ -73,7 +73,7 @@ class AttendanceFactory extends Factory
         return $this->state(function (array $attributes) {
             $timeIn = $this->faker->dateTimeBetween('-1 month', '-1 day');
             $timeOut = Carbon::parse($timeIn)->addMinutes($this->faker->numberBetween(60, 480));
-            $durationMinutes = Carbon::parse($timeIn)->diffInMinutes($timeOut);
+            $durationMinutes = (int) Carbon::parse($timeIn)->diffInMinutes($timeOut);
 
             return [
                 'time_in' => $timeIn,
